requirements.txt
torch>=2.0.0
transformers>=4.38.0
datasets>=2.18.0
peft>=0.10.0
evaluate>=0.4.1
accelerate>=0.27.0
wandb>=0.16.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
scikit-learn>=1.3.0
pandas>=2.0.0
rouge-score>=0.1.2
sacrebleu>=2.3.1
sentencepiece>=0.1.99
